requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
google-genai
pymupdf
python-multipart
pydantic
python-dotenv
